move signup function to user.ts
