+++ lib/db/user.ts
@@ -1,6 +1,7 @@
 import prisma from "./prisma";
 import { User } from "@prisma/client";
 import { UserWithIncludes } from "./types";
+import { genSaltSync, hashSync } from "bcrypt-ts/browser";
 
 const inclusions = {
   task: true,
@@ -82,5 +83,20 @@ export async function deleteUser(id: number): Promise<User> {
   const user = await prisma.user.delete({
     where: { id: id },
   });
+  return user;
+}
+
+/**
+ * Signs a user up for a lookout account.
+ *
+ * @param {Omit<User, "id | password | salt">} params - User details
+ * @returns {Promise<User>} - New user
+ */
+export async function signUp(
+  params: Omit<User, "id | password | salt">
+): Promise<User> {
+  params.salt = genSaltSync(10);
+  params.password = hashSync(params.password, params.salt);
+  const user = await createUser(params);
   return user;
 }
